Put primary domain type definitions at package root
They tend to be used pervasively throughout the package, so this is
usually convenient.
